packages/services/health-costs-api/requirements.txt
uvicorn
fastapi
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cssselect==1.1.0
cssutils==1.0.2
dnspython==1.16.0
email-validator==1.0.5
emails==0.5.15
fastapi==0.45.0
h11==0.8.1
idna==2.8
lxml==4.4.2
passlib==1.7.2
premailer==3.6.1
psycopg2-binary==2.8.4
pydantic==1.2
PyJWT==1.7.1
python-dateutil==2.8.1
python-multipart==0.0.5
requests==2.22.0
six==1.13.0
SQLAlchemy==1.3.12
starlette==0.12.9
urllib3==1.25.7
uvicorn==0.11.7
websockets==8.1
